Vessel list

+++ src/main/java/com/shipping/dao/vessel/VesselOwnerRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 import com.shipping.model.crew.Crew;
 import com.shipping.model.vessel.VesselOwner;
 
-public interface VesselOwnerRepository extends MongoRepository<VesselOwner, Integer> {
+public interface VesselOwnerRepository extends MongoRepository<VesselOwner, Long> {
 	//public Crew save(Crew crew);
 
 	//public Crew findByFirstName(String firstName);
